backend/requirements.txt
fastapi
uvicorn
motor
pydantic
google-generativeai
python-dotenv
python-multipart
pillow
